# openapi.json
{
  "components": {
    "schemas": {
...
      "AiProviderConfigCreate": {
        "description": "Create a new AI provider configuration.",
        "properties": {
          "api_key": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "default_model": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Default Model",
            "type": "string"
          },
          "is_enabled": {
            "default": true,
            "title": "Is Enabled",
            "type": "boolean"
          },
          "label": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "provider": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "label",
          "api_key",
          "default_model"
        ],
        "title": "AiProviderConfigCreate",
        "type": "object"
      },
      "AiProviderConfigRead": {
        "description": "AI provider configuration response.",
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_model": {
            "title": "Default Model",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_enabled": {
            "title": "Is Enabled",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider",
          "label",
          "api_key",
          "base_url",
          "default_model",
          "is_enabled",
          "created_at",
          "updated_at"
        ],
        "title": "AiProviderConfigRead",
        "type": "object"
      },
      "AiProviderConfigUpdate": {
        "description": "Update an AI provider configuration.",
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "default_model": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model"
          },
          "is_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Enabled"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "title": "AiProviderConfigUpdate",
        "type": "object"
      },
      "AiSuggestionCreate": {
        "description": "AI suggestion request.",
        "properties": {
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "AiSuggestionCreate",
        "type": "object"
      },
      "AiSuggestionRead": {
        "description": "AI suggestion response.",
        "properties": {
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "suggestion": {
            "title": "Suggestion",
            "type": "string"
          }
        },
        "required": [
          "id",
          "prompt",
          "context",
          "suggestion",
          "created_at"
        ],
        "title": "AiSuggestionRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngredientSubstitutionRequest": {
        "properties": {
          "expiry_items": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Expiry Items",
            "type": "array"
          },
          "inventory_items": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Inventory Items",
            "type": "array"
          },
          "recipe_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Recipe Ingredients",
            "type": "array"
          },
          "recipe_title": {
            "title": "Recipe Title",
            "type": "string"
          }
        },
        "required": [
          "recipe_title"
        ],
        "title": "IngredientSubstitutionRequest",
        "type": "object"
      },
...
      "PhotoAnalysisRequest": {
        "properties": {
          "image_paths": {
            "items": {
              "type": "string"
            },
            "title": "Image Paths",
            "type": "array"
          }
        },
        "title": "PhotoAnalysisRequest",
        "type": "object"
      },
...
      "RecipeEnhanceRequest": {
        "properties": {
          "recipe_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Recipe Ingredients",
            "type": "array"
          },
          "recipe_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Instructions"
          },
          "recipe_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Summary"
          },
          "recipe_title": {
            "title": "Recipe Title",
            "type": "string"
          }
        },
        "required": [
          "recipe_title"
        ],
        "title": "RecipeEnhanceRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Self-hostable kitchen OS API for recipes, inventory, shopping, scanners, and AI-ready workflows.",
    "title": "Kombu",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/inventory/analyze-photos": {
      "post": {
        "description": "Analyze food photos to identify inventory items.",
        "operationId": "analyze_photos_api_v1_ai_inventory_analyze_photos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Analyze Photos Api V1 Ai Inventory Analyze Photos Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Photos",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/providers": {
      "get": {
        "description": "List all configured AI providers.",
        "operationId": "list_providers_api_v1_ai_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AiProviderConfigRead"
                  },
                  "title": "Response List Providers Api V1 Ai Providers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Providers",
        "tags": [
          "ai"
        ]
      },
      "post": {
        "description": "Add a new AI provider configuration.",
        "operationId": "create_provider_api_v1_ai_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiProviderConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderConfigRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Provider",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/v1/ai/providers/{provider_id}": {
      "delete": {
        "description": "Delete an AI provider configuration.",
        "operationId": "delete_provider_api_v1_ai_providers__provider_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Provider",
        "tags": [
          "ai"
        ]
      },
      "patch": {
        "description": "Update an AI provider configuration.",
        "operationId": "update_provider_api_v1_ai_providers__provider_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiProviderConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderConfigRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Provider",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/recipes/enhance": {
      "post": {
        "description": "Enhance a recipe with AI-powered formatting.",
        "operationId": "enhance_recipe_api_v1_ai_recipes_enhance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeEnhanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enhance Recipe Api V1 Ai Recipes Enhance Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enhance Recipe",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/recipes/substitutions": {
      "post": {
        "description": "Get AI-powered ingredient substitution suggestions.",
        "operationId": "ingredient_substitutions_api_v1_ai_recipes_substitutions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientSubstitutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ingredient Substitutions Api V1 Ai Recipes Substitutions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingredient Substitutions",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/v1/ai/suggestions": {
      "post": {
        "description": "Create a provider-free AI suggestion record.",
        "operationId": "suggest_api_v1_ai_suggestions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiSuggestionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSuggestionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}